README.md
An Artificial Neural Network (ANN) Model for an Effluent Treatment Plant (ETP) is a computational framework designed to predict and optimize various aspects of the ETP's performance. An ETP is a facility used to treat industrial wastewater before it is discharged into the environment to minimize its negative impact on nature and public health. The ANN model can be a valuable tool in enhancing the efficiency and effectiveness of ETP operations.